Revert "Moved Listener to its own place. Created res folder to store all this…"
