+++ src/index.js
@@ -1,9 +1,14 @@
 import Flask from './Flask'
-import * as listeners from './res/listeners'
+import {
+  GLOBAL_LISTENER_NAME,
+  RESOLVED_LISTENER_NAME
+} from './util/constants'
 
 Flask._version = '1.0.3'
 Flask._env = 'development'
 
-Flask.listeners = listeners
-
 export default Flask
+export {
+  GLOBAL_LISTENER_NAME,
+  RESOLVED_LISTENER_NAME
+}
