--- src/res/listeners.js
@@ -1,2 +1,0 @@
-export const GLOBAL_LISTENER_NAME = '__global__'
-export const ON_RESOLVED = 'resolved'
